Moved local s3 utility code to automation.utilities
